Avoid consolidate package import cycles
