+++ python-avd/pyavd/_eos_designs/consolidate/__init__.py
@@ -1,7 +1,3 @@
 # Copyright (c) 2026 Arista Networks, Inc.
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
-from .model import ConsolidatedAVDDesign, PrunedAVDDesign
-from .models import ConsolidatedData
-
-__all__ = ["ConsolidatedAVDDesign", "ConsolidatedData", "PrunedAVDDesign"]
