+++ python-avd/pyavd/_eos_designs/structured_config/structured_config_generator.py
@@ -18,7 +18,7 @@ if TYPE_CHECKING:
 
     from typing_extensions import Self
 
-    from pyavd._eos_designs.consolidate import PrunedAVDDesign
+    from pyavd._eos_designs.consolidate.model import PrunedAVDDesign
     from pyavd._eos_designs.eos_designs_facts.schema import EosDesignsFacts
     from pyavd._eos_designs.shared_utils import SharedUtilsProtocol
     from pyavd._eos_designs.structured_config.structured_config_utils import StructuredConfigUtils
