+++ python-avd/pyavd/api/schemas/__init__.py
@@ -3,7 +3,7 @@
 # that can be found in the LICENSE file.
 from pyavd._cv.schema import CvDeploy as CVDeploy
 from pyavd._eos_cli_config_gen.schema import EosCliConfigGen as EOSConfig
-from pyavd._eos_designs.consolidate import ConsolidatedAVDDesign
+from pyavd._eos_designs.consolidate.model import ConsolidatedAVDDesign
 from pyavd._eos_designs.schema import EosDesigns as AVDDesign
 
 __all__ = ["AVDDesign", "CVDeploy", "ConsolidatedAVDDesign", "EOSConfig"]
