+++ python-avd/tests/models.py
@@ -14,7 +14,7 @@ from ansible.parsing.dataloader import DataLoader
 from ansible.vars.manager import VariableManager
 from yaml import CSafeLoader, load
 
-from pyavd._eos_designs.consolidate import ConsolidatedAVDDesign
+from pyavd._eos_designs.consolidate.model import ConsolidatedAVDDesign
 from pyavd._eos_designs.eos_designs_facts.get_facts import get_facts
 from pyavd._utils import get
 from pyavd.api.pool_manager import PoolManager
